Used arrays for executing the query in national parks table

+++ park_migration.php
@@ -17,5 +17,6 @@ $dbc->exec('CREATE TABLE national_parks(
 	name VARCHAR(255) NOT NULL,
 	location VARCHAR(255) NOT NULL,
 	date_established DATE NOT NULL,
-	area_in_acres FLOAT NOT NULL
+	area_in_acres FLOAT(12,2) NOT NULL
 )');
+
